Feat: 디렉토리 유지보수
resolve #265
