+++ backend/src/main/java/com/example/nalot/controller/WeatherController.java
@@ -71,55 +71,4 @@ public class WeatherController {
         return weatherService.setWeatherDto(items,dateNow.get(0),dateNow.get(1));
     }
 
-    //@EventListener(ApplicationReadyEvent.class)
-    @GetMapping("/data")
-    public Dataset<Row> getWeatherDataset() {
-        Dataset<Row> df = weatherService.getWeatherDataset();
-        df.printSchema();
-        return df;
-    }
-
-    //@EventListener(ApplicationReadyEvent.class)
-    @GetMapping("/refine")
-    public Dataset<Row> refineDataSet() {
-        Dataset<Row> df = weatherService.getWeatherDataset();
-        return weatherService.refineDataSet(df);
-
-    }
-
-    //@EventListener(ApplicationReadyEvent.class)
-    public Dataset<Row> getLocationDataset(){
-        Dataset<Row> df = weatherService.getWeatherDataset();
-        Dataset<Row> ds =  weatherService.refineDataSet(df);
-        Dataset<Row> ds2 = weatherService.getLocationDataset(ds);
-
-        ds2.sort(ds.col("date").desc()).show();
-
-        return ds2;
-    }
-
-    //@EventListener(ApplicationReadyEvent.class)
-    public Dataset<Row> refineClothesDataset(){
-        Dataset<Row> df = weatherService.getClothesDataset();
-        Dataset<Row> ds = weatherService.refineClothesData(df);
-
-        return ds;
-    }
-
-    @EventListener(ApplicationReadyEvent.class)
-    public Dataset<Row> joinDataSet(){
-        Dataset<Row> df = weatherService.getWeatherDataset();
-        Dataset<Row> ds =  weatherService.refineDataSet(df);
-        Dataset<Row> temp = weatherService.getLocationDataset(ds);
-        //temp.sort(ds.col("date").desc()).show();
-
-        Dataset<Row> df2 = weatherService.getClothesDataset();
-        Dataset<Row> clothes = weatherService.refineClothesData(df2);
-
-        temp.show();
-        clothes.show();
-        Dataset<Row> result = weatherService.joinDataSet(temp,clothes);
-
-        return result;
-    }
 }
